Flows

+++ static/components/documentComponent/document.js
@@ -22,6 +22,8 @@ export default class DocumentComponent extends AbstractComponent {
             "format_open": this.$element.hasClass("document--format")
         }
         // Others
+        this.reply;
+        this.editorJS = false;
         this.editor;
         this.editorData;
         this.task = task;
@@ -33,21 +35,23 @@ export default class DocumentComponent extends AbstractComponent {
 
     init() {
         // Formats Kebab
-        this.kebabFormats = new KebabComponent(this.$headerFormats.find("#document-formats-kebab"));
-        this.getFormats().then((formats) => {;
-            for (const format of formats) {
-                let $item = this.kebabFormats.addItem(format.name, format.id);
-                $item.on("click", (e) => { 
-                    console.log(format.id);    
-                    this.setState({
-                        "status":"loading",
-                        "id": format.id,
-                        "format_open": true
-                    });
-                    this.editorLoad();
-                })
-            }
-        })
+        if (this.$header.length != 0) {
+            this.kebabFormats = new KebabComponent(this.$headerFormats.find("#document-formats-kebab"));
+            this.getFormats().then((formats) => {;
+                for (const format of formats) {
+                    let $item = this.kebabFormats.addItem(format.name, format.id);
+                    $item.on("click", (e) => { 
+                        console.log(format.id);    
+                        this.setState({
+                            "status":"loading",
+                            "id": format.id,
+                            "format_open": true
+                        });
+                        this.editorLoad();
+                    })
+                }
+            })
+        }
 
         // Formats selector
         /*
@@ -84,9 +88,8 @@ export default class DocumentComponent extends AbstractComponent {
             this.save(this.state["id"]);
         })
 
-        // Read the document, if there is id.
+        // Load the editor.
         if (this.state.id) {
-            //this.read(this.state.id);
             this.editorLoad();
         }
 
@@ -101,12 +104,9 @@ export default class DocumentComponent extends AbstractComponent {
                 type: 'GET',
                 data: {"is_format": true, "parent_document_id": this.state["id"]},
                 success: (data) => {
-                    console.log('Formats: ', data)
                     resolve(data);
                 },
                 error: (jqXHR, textStatus, errorThrown) => {
-                    console.log('Error: ', textStatus)
-                    console.log('Get formats failed: ', errorThrown)
                     reject();
                 }
             
@@ -206,7 +206,6 @@ export default class DocumentComponent extends AbstractComponent {
         this.editorLoad();
 
         let url = `/document/${this.state["id"]}/`
-        
         // Call document:read
         return new Promise((resolve, reject) => { 
             jQuery.ajax({
@@ -225,7 +224,6 @@ export default class DocumentComponent extends AbstractComponent {
                     if (jQuery(".document").hasClass("document--format")) {
                         this.openFormat();
                     }
-                    console.log("DocumentComponent.read finished", this.state["id"]);
                     this.setState({"status": "idle"});
                     resolve();
                 }
@@ -235,48 +233,63 @@ export default class DocumentComponent extends AbstractComponent {
 
     editorLoad() {
         this.$element.find("#document-editor").html("");
-        this.editor = new EditorJS({
-            holder: 'document-editor',
-            tools: {
-                header: Header,
-                delimiter: Delimiter,
-                paragraph: {
-                    class: Paragraph,
-                    inlineToolbar: true,
-                },
-                list: {
-                    class: List,
-                    inlineToolbar: true,
-                    config: {
-                        defaultStyle: 'unordered'
+        if (this.editorJS) {
+            this.editor = new EditorJS({
+                holder: 'document-editor',
+                tools: {
+                    header: Header,
+                    delimiter: Delimiter,
+                    paragraph: {
+                        class: Paragraph,
+                        inlineToolbar: true,
+                    },
+                    list: {
+                        class: List,
+                        inlineToolbar: true,
+                        config: {
+                            defaultStyle: 'unordered'
+                        }
                     }
+                },
+                onReady: () => {
+                    console.log('Editor.js is ready to work!')
+                    this.editorRead();
                 }
-            },
-            onReady: () => {
-                console.log('Editor.js is ready to work!')
-                this.editorRead();
-            }
-        });
+            });
+        } else {
+            this.editorRead();
+        };
     }
 
     editorRead() {
         jQuery("#document-editor-wrapper-indicator").show();
         jQuery("#document-editor-wrapper").hide();
         jQuery.ajax({
-            url: `/document/api/v1/document/${this.state["id"]}`,
+            url: `/document/api/v1/document/${this.state["id"]}/reply`,
             type: 'GET',
             success: (data) => {
-                console.log("data",data);
                 jQuery("#document-editor-wrapper-indicator").hide();
                 jQuery("#document-editor-wrapper").show();
+                console.log(data);
                 // Update if data is not empty
                 if (!jQuery.isEmptyObject(data)) {
                     this.setState({"status":"idle"});
-                    this.editorData = data;
-                    console.log("Caca",this.editorData)
-                    this.editorUpdate();
-                    if (this.task) {
-                        this.task.documentReady();
+                    if (this.editorJS) {
+                        this.editorData = data;
+                        this.editorUpdate();
+                        if (this.task) {
+                            //this.task.documentReady();
+                        }
+                    } else {
+                        if (data.reply) {
+                            this.reply = data.reply;
+                            let md = markdownit();
+                            let html = md.render(this.reply);
+                            jQuery("#document-editor").html(html);
+                        } else {
+                            jQuery("#document-editor").html("");
+                        }
+                        
                     }
                 }
             }
@@ -284,28 +297,41 @@ export default class DocumentComponent extends AbstractComponent {
     }
 
     editorClear() {
-        console.log("DocumentComponent.editorClear");
-        this.editor.clear();
+        if (this.editorJS) {
+            console.log("DocumentComponent.editorClear");
+            if (this.editor) {
+                this.editor.clear();
+            } else {
+                console.warn("DocumentComponent.editorClear: editor not found")
+            }
+        } else {
+            this.$element.find("#document-editor").html("");
+        }
     }
 
     editorUpdate(save = false) {
-        console.log("DocumentComponent.editorUpdate()", save)
-        this.editor.clear().then(() => {
-            console.log("DocumentComponent.editorUpdate():clear")
-            this.editor.render(this.editorData).then(() => {
-                console.log("DocumentComponent.editorUpdate():rendered")
-                this.editor.save().then((output_data) => {
-                    console.log("DocumentComponent.editorUpdate():saved")
-                    this.editorData = output_data;
-                    this.loadSections();
-                    if (save) {
-                        this.save_back(this.state["id"], output_data)
-                    }
-                }).catch((error) => {
-                    console.log('# Saving document failed: ', error)
+        if (this.editorJS) {
+            this.editor.clear().then(() => {
+                this.editor.render(this.editorData).then(() => {
+                    this.editor.save().then((output_data) => {
+                        this.editorData = output_data;
+                        this.loadSections();
+                        if (save) {
+                            this.save_back(this.state["id"], output_data)
+                        }
+                    }).catch((error) => {
+                        console.log('# Saving document failed: ', error)
+                    });
                 });
-            });
-        })
+            })
+        } else {
+            let md = markdownit();
+            let result = md.render(this.reply);
+            jQuery("#document-editor").html(result);
+            if (save) {
+                this.save_back(this.state["id"], this.reply, false)
+            }
+        }
     }
 
     stream(instruction, instruction_element_id) {
@@ -345,8 +371,6 @@ export default class DocumentComponent extends AbstractComponent {
                 get_prompt_data[jQuery(this).attr("name")] = jQuery(this).attr("value")
             }
         });
-
-        console.log("get_prompt_data", get_prompt_data);
     
         // Call to inititate stream. Get the prompt and then call the LLM.
         jQuery.ajax({
@@ -378,18 +402,19 @@ export default class DocumentComponent extends AbstractComponent {
                 // Only update after 1 second
                 if (time_since_last_message > 1000) {
                     // Update the document with the response
+                    this.reply = data.response_text;
                     this.editorData = data.document_json;
                     this.editorUpdate()
                     time_last_message = Date.now();
+
                 }
-                //document_update(document_data)
-    
             } else if (data.status == "end") {
                 // Stream finished, update the document
-                console.log("Stream end", instruction);
+                console.log("Stream end", instruction);    
                 this.editorData = data.document_json;
+                this.reply = data.response_text;
                 this.editorUpdate(true);
-                this.task.documentReady();
+                //this.task.documentReady();
                 instruction.setState({"status":"idle"});
                 this.setState({"status": "idle"})
                 socket.close();
@@ -403,17 +428,26 @@ export default class DocumentComponent extends AbstractComponent {
     
     }
 
-    save_back(document_id, output_data) {
+    save_back(document_id, output_data, json=true) {
         console.log("document_save_back", document_id);
         let csrfToken = Cookies.get('csrftoken');
         let url = `/document/api/v1/document/${document_id}/save`
-        console.log(url);
+        let document_reply = null;
+        let document_json = null;
+        if (json) {
+            document_json = JSON.stringify(output_data);
+        } else {
+            document_reply = output_data;
+        }
+        console.log(document_json);
+        console.log(document_reply);
         jQuery.ajax({
             url: url,
             type: 'POST',
             data: {
                 'csrfmiddlewaretoken': csrfToken,
-                'document_data': JSON.stringify(output_data)
+                'document_reply': document_reply,
+                'document_json': document_json
             },
             dataType: 'json',
             success: (data) => {
@@ -429,12 +463,16 @@ export default class DocumentComponent extends AbstractComponent {
     
     save(document_id) {
         console.log("save", document_id)
-        this.editor.save().then((output_data) => {
-            console.log('Document data: ', output_data)
-            this.save_back(document_id, output_data)
-        }).catch((error) => {
-            console.log('Saving document failed: ', error)
-        });
+        if (!this.editorJS) {
+            this.save_back(document_id, this.reply, false);
+        } else {
+            this.editor.save().then((output_data) => {
+                console.log('Document data: ', output_data)
+                this.save_back(document_id, output_data)
+            }).catch((error) => {
+                console.log('Saving document failed: ', error)
+            });
+        }
     }
 
     /* Editor editing */
